validate_uncertainty [skip hazardlib]
