+++ openquake/commonlib/lt.py
@@ -16,10 +16,37 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with OpenQuake.  If not, see <http://www.gnu.org/licenses/>.
 
+import re
+
 from openquake.baselib.general import CallableDict
 from openquake.hazardlib import geo
 from openquake.hazardlib.sourceconverter import (
     split_coords_2d, split_coords_3d)
+
+
+class LogicTreeError(Exception):
+    """
+    Logic tree file contains a logic error.
+
+    :param node:
+        XML node object that causes fail. Used to determine
+        the affected line number.
+
+    All other constructor parameters are passed to :class:`superclass'
+    <LogicTreeError>` constructor.
+    """
+    def __init__(self, node, filename, message):
+        self.filename = filename
+        self.message = message
+        self.lineno = node if isinstance(node, int) else getattr(
+            node, 'lineno', '?')
+
+    def __str__(self):
+        return "filename '%s', line %s: %s" % (
+            self.filename, self.lineno, self.message)
+
+
+#                           parse_uncertainty                              #
 
 parse_uncertainty = CallableDict(
     keymissing=lambda utype, node: float(node.text.strip()))
@@ -50,7 +77,6 @@ def simpleGeom(utype, node):
     spacing = node["spacing"]
     usd, lsd, dip = (~node.upperSeismoDepth, ~node.lowerSeismoDepth,
                      ~node.dip)
-    # Parse the geometry
     coords = split_coords_2d(~node.LineString.posList)
     trace = geo.Line([geo.Point(*p) for p in coords])
     return trace, usd, lsd, dip, spacing
@@ -98,3 +124,108 @@ def charGeom(utype, node):
         return geo.MultiSurface(surfaces)
     else:
         return surfaces[0]
+
+
+#                             validate_uncertainty                          #
+
+_float_re = re.compile(r'^(\+|\-)?(\d+|\d*\.\d+)$')
+
+
+def unknown(utype, node, filename):
+    try:
+        float(node.text)
+    except (TypeError, ValueError):
+        raise LogicTreeError(node, filename, 'expected single float value')
+
+
+validate_uncertainty = CallableDict(keymissing=unknown)
+
+
+@validate_uncertainty.add('simpleFaultGeometryAbsolute')
+def _validate_simple_fault_geometry(utype, node, filename):
+    if hasattr(node, 'simpleFaultGeometry'):
+        node = node.simpleFaultGeometry
+    try:
+        coords = split_coords_2d(~node.LineString.posList)
+        trace = geo.Line([geo.Point(*p) for p in coords])
+    except ValueError:
+        # If the geometry cannot be created then use the LogicTreeError
+        # to point the user to the incorrect node. Hence, if trace is
+        # compiled successfully then len(trace) is True, otherwise it is
+        # False
+        trace = []
+    if len(trace):
+        return
+    raise LogicTreeError(
+        node, filename, "'simpleFaultGeometry' node is not valid")
+
+
+@validate_uncertainty.add('complexFaultGeometryAbsolute')
+def _validate_complex_fault_geometry(utype, node, filename):
+    # NB: if the geometry does not conform to the Aki & Richards convention
+    # this will not be verified here, but will raise an error when the surface
+    # is created
+    if hasattr(node, 'complexFaultGeometry'):
+        node = node.complexFaultGeometry
+    valid_edges = []
+    for edge_node in node.nodes:
+        try:
+            coords = split_coords_3d(edge_node.LineString.posList.text)
+            edge = geo.Line([geo.Point(*p) for p in coords])
+        except ValueError:
+            # See use of validation error in simple geometry case
+            # The node is valid if all of the edges compile correctly
+            edge = []
+        if len(edge):
+            valid_edges.append(True)
+        else:
+            valid_edges.append(False)
+    if node["spacing"] and all(valid_edges):
+        return
+    raise LogicTreeError(
+        node, filename, "'complexFaultGeometry' node is not valid")
+
+
+@validate_uncertainty.add('planarSurface')
+def _validate_planar_fault_geometry(utype, node, filename):
+    valid_spacing = node["spacing"]
+    for key in ["topLeft", "topRight", "bottomLeft", "bottomRight"]:
+        lon = getattr(node, key)["lon"]
+        lat = getattr(node, key)["lat"]
+        depth = getattr(node, key)["depth"]
+        valid_lon = (lon >= -180.0) and (lon <= 180.0)
+        valid_lat = (lat >= -90.0) and (lat <= 90.0)
+        valid_depth = (depth >= 0.0)
+        is_valid = valid_lon and valid_lat and valid_depth
+        if not is_valid or not valid_spacing:
+            raise LogicTreeError(
+                node, filename, "'planarFaultGeometry' node is not valid")
+
+
+@validate_uncertainty.add('characteristicFaultGeometryAbsolute')
+def _validate_characteristic_fault_geometry(utype, node, filename):
+    for geom_node in node.surface:
+        if "simpleFaultGeometry" in geom_node.tag:
+            _validate_simple_fault_geometry(utype, geom_node, filename)
+        elif "complexFaultGeometry" in geom_node.tag:
+            _validate_complex_fault_geometry(utype, geom_node, filename)
+        elif "planarSurface" in geom_node.tag:
+            _validate_planar_fault_geometry(utype, geom_node, filename)
+        else:
+            raise LogicTreeError(
+                geom_node, filename, "Surface geometry type not recognised")
+
+
+@validate_uncertainty.add('abGRAbsolute')
+def _validate_abGRAbsolute(utype, node, filename):
+    ab = node.text.strip().split()
+    if len(ab) == 2:
+        if _float_re.match(ab[0]) and _float_re.match(ab[1]):
+            return
+    raise LogicTreeError(
+        node, filename, 'expected a pair of floats separated by space')
+
+
+@validate_uncertainty.add('incrementalMFDAbsolute')
+def _validate_incMFD(utype, node, filename):
+    pass
